apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic[email]==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
passlib==1.7.4
python-jose==3.5.0
reportlab==4.4.3
python-docx==1.2.0
python-slugify==8.0.4
